balancing commands file made :^)
